Move sectorinterface into vfs.
